+++ Optimizer/Controls/MoonMenuRenderer.cs
@@ -1,5 +1,4 @@
-﻿using Optimizer;
-using System.Drawing;
+﻿using System.Drawing;
 using System.Windows.Forms;
 
 namespace Optimizer