+++ Optimizer/PingerHelper.cs
@@ -158,7 +158,7 @@ namespace Optimizer
         //            return $"{port} - [×]";
         //        }
 
-                
+
         //    }
         //    catch
         //    {